# docs/swagger.yaml
definitions:
...
  model.Log:
    properties:
      address:
        description: address
        type: string
      data:
        description: data
        type: string
      logIndex:
        description: Transaction random number, transaction volume
        type: integer
      removed:
        description: whether to remove
        type: boolean
      topics:
        description: Uncle block hash
        items:
          type: string
        type: array
      transactionHash:
        description: Create transaction
        type: string
    type: object
...
  model.Reward:
    properties:
      address:
        description: reward address
        type: string
      amount:
        description: Amount of reward
        type: string
      block_number:
        description: The block number when rewarding
        type: integer
      identity:
        description: 'Identity, 1: block producer, 2: verifier, 3, exchange'
        type: integer
      snft:
        description: SNFT address
        type: string
    type: object
...
  model.Transaction:
    properties:
      blockHash:
        description: Create transaction
        type: string
      blockNumber:
        description: Transaction random number, transaction volume
        type: integer
      contractAddress:
        description: address
        type: string
      cumulativeGasUsed:
        description: Transaction random number, transaction volume
        type: integer
      from:
        description: address
        type: string
      gas:
        description: Transaction random number, transaction volume
        type: integer
      gasPrice:
        description: Transaction random number, transaction volume
        type: integer
      gasUsed:
        description: Transaction random number, transaction volume
        type: integer
      hash:
        description: Create transaction
        type: string
      input:
        description: Additional input data, contract call encoded data
        type: string
      methodId:
        description: difficulty random number
        type: string
      nonce:
        description: Transaction random number, transaction volume
        type: integer
      status:
        description: Transaction random number, transaction volume
        type: integer
      to:
        description: address
        type: string
      transactionIndex:
        description: Transaction random number, transaction volume
        type: integer
      value:
        description: The total amount of coins in the chain
        type: string
    type: object
...
  service.ErrRes:
    properties:
      err_str:
        description: Error message
        type: string
    type: object
...
  service.RankingExchangerRes:
    properties:
      exchangers:
        description: Exchanger list
        items:
          properties:
            address:
              description: Exchange address
              type: string
            balance_count:
              description: Total transaction amount, unit wei
              type: string
            block_number:
              description: The block number when created
              type: integer
            creator:
              description: Creator address
              type: string
            fee_ratio:
              description: fee rate, unit 1/10,000
              type: integer
            name:
              description: Exchange name
              type: string
            nft_count:
              description: Total NFT count
              type: integer
            timestamp:
              description: Open time
              type: integer
            tx_hash:
              description: The transaction created
              type: string
            txCount:
              type: integer
            url:
              description: Exchange URL
              type: string
          type: object
        type: array
      total:
        description: The total number of Exchanger
        type: integer
    type: object
  service.RankingNFTRes:
    properties:
      nfts:
        description: NFT list
        items:
          properties:
            address:
              description: NFT address, grows automatically from 0x1
              type: string
            attributes:
              description: Attributes
              type: string
            block_number:
              description: The height of the created block
              type: integer
            category:
              description: category
              type: string
            collection_id:
              description: The id of the collection, the name of the collection +
                the creator of the collection + the hash of the exchange where the
                collection is located
              type: string
            creator:
              description: Creator address
              type: string
            desc:
              description: description
              type: string
            exchanger_addr:
              description: The address of the exchange, if there is none, it can be
                traded on any exchange
              type: string
            last_price:
              description: The last transaction price (null if the transaction is
                not completed), the unit is wei
              type: string
            meta_url:
              description: Real meta information URL
              type: string
            name:
              description: name
              type: string
            owner:
              description: owner
              type: string
            raw_meta_url:
              description: Original meta information URL on the chain
              type: string
            royalty_ratio:
              description: Royalty rate, in ten thousandths
              type: integer
            source_url:
              description: Resource links, file links such as pictures or videos
              type: string
            timestamp:
              description: Create timestamp
              type: integer
            tx_hash:
              description: The transaction hash created
              type: string
            txCount:
              type: integer
          type: object
        type: array
      total:
        description: The total number of NFTs
        type: integer
    type: object
  service.RankingSNFTRes:
    properties:
      nfts:
        description: SNFT list
        items:
          properties:
            address:
              description: SNFT address
              type: string
            awardee:
              description: The address of the miner that was rewarded last, null if
                it has not been rewarded
              type: string
            creator:
              type: string
            last_price:
              description: The last transaction price, the unit is wei, null if the
                transaction has not been completed
              type: string
            owner:
              description: owner, unallocated and reclaimed are null
              type: string
            reward_at:
              description: The timestamp of the last rewarded, null if not rewarded
              type: integer
            reward_number:
              description: The height of the last rewarded block, null if not rewarded
              type: integer
            txCount:
              type: integer
          type: object
        type: array
      total:
        description: The total number of SNFTs
        type: integer
    type: object
  service.RewardsRes:
    properties:
      rewards:
        description: Rewards list
        items:
          $ref: '#/definitions/model.Reward'
        type: array
      total:
        description: The total number of rewards
        type: integer
    type: object
...
  service.TransactionsRes:
    properties:
      total:
        description: The total number of transactions
        type: integer
      transactions:
        description: Transaction list
        items:
          properties:
            blockHash:
              description: Create transaction
              type: string
            blockNumber:
              description: Transaction random number, transaction volume
              type: integer
            contractAddress:
              description: address
              type: string
            cumulativeGasUsed:
              description: Transaction random number, transaction volume
              type: integer
            from:
              description: address
              type: string
            gas:
              description: Transaction random number, transaction volume
              type: integer
            gasPrice:
              description: Transaction random number, transaction volume
              type: integer
            gasUsed:
              description: Transaction random number, transaction volume
              type: integer
            hash:
              description: Create transaction
              type: string
            input:
              description: Additional input data, contract call encoded data
              type: string
            methodId:
              description: difficulty random number
              type: string
            nonce:
              description: Transaction random number, transaction volume
              type: integer
            status:
              description: Transaction random number, transaction volume
              type: integer
            timestamp:
              description: The event stamp of the block it is in
              type: integer
            to:
              description: address
              type: string
            transactionIndex:
              description: Transaction random number, transaction volume
              type: integer
            value:
              description: The total amount of coins in the chain
              type: string
          type: object
        type: array
    type: object
info:
  contact: {}
  description: Block browser back-end interface, parses data from the blockchain,
    provides information retrieval services for blocks, transactions, NFT, SNFT, NFT
    collections, and exchanges
  title: block explorer API
  version: "1.0"
paths:
...
  /ranking/exchanger:
    get:
      consumes:
      - application/json
      description: Exchanger ranking for a specified time range
      parameters:
      - description: Limit, time range
        in: query
        name: limit
        type: string
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.RankingExchangerRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query exchanger ranking
      tags:
      - Ranking
  /ranking/nft:
    get:
      consumes:
      - application/json
      description: NFT ranking for a specified time range
      parameters:
      - description: Limit, time range
        in: query
        name: limit
        type: string
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.RankingNFTRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query NFT ranking
      tags:
      - Ranking
  /ranking/snft:
    get:
      consumes:
      - application/json
      description: SNFT ranking for a specified time range
      parameters:
      - description: Limit, time range
        in: query
        name: limit
        type: string
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.RankingSNFTRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query SNFT ranking
      tags:
      - Ranking
  /reward:
    get:
      consumes:
      - application/json
      description: query the reward list in reverse order
      parameters:
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.RewardsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: Query the reward list
      tags:
      - reward
  /reward/{block}:
    get:
      consumes:
      - application/json
      description: specifies the block query reward
      parameters:
      - description: Block height
        in: path
        name: block
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Reward'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query reward
      tags:
      - reward
...
  /transaction/{hash}:
    get:
      consumes:
      - application/json
      description: specifies the hash query transaction
      parameters:
      - description: Transaction hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query transaction
      tags:
      - transaction
  /transaction/page:
    get:
      consumes:
      - application/json
      description: query transaction list in reverse order
      parameters:
      - description: Page, default 1
        in: query
        name: page
        type: string
      - description: Page size, default 10
        in: query
        name: page_size
        type: string
      - description: Block number, if empty, query all
        in: query
        name: number
        type: string
      - description: Account address, if empty, query all
        in: query
        name: addr
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TransactionsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query transaction list
      tags:
      - transaction
  /transaction_logs/{hash}:
    get:
      consumes:
      - application/json
      description: specifies the transaction hash to query the transaction receipt
      parameters:
      - description: Transaction hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Log'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/service.ErrRes'
      summary: query transaction receipt
      tags:
      - transaction
swagger: "2.0"